# Feature/speedup ascii