Fix #170 

results from local testing:
std::format: 5.3753 seconds
std::ostringstream: 8.68944 seconds
before: 9.99393 seconds
dune: 12.4421 seconds